+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.5.0 (2019-08-10)
 
 * **Breaking change:** Now requires Node.js >=10.x
+* **Dependencies:** Updated to `promised-retry` version `0.3.x`
+* **Improvement:** Improved JSDoc documentation + added linting of it through TypeScript
 
 ## 0.4.1 (2018-09-17)